requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
mysql-connector-python
flask-cors